change to use boolean flag, so dummy_ecdsa method lives in ecdsa
