+++ cpp/src/barretenberg/dsl/acir_format/ecdsa_secp256k1.cpp
@@ -84,8 +84,14 @@ witness_ct ecdsa_index_to_witness(Composer& composer, uint32_t index)
     return { &composer, value };
 }
 
-void create_ecdsa_verify_constraints(Composer& composer, const EcdsaSecp256k1Constraint& input)
+void create_ecdsa_verify_constraints(Composer& composer,
+                                     const EcdsaSecp256k1Constraint& input,
+                                     bool has_valid_witness_assignments)
 {
+
+    if (has_valid_witness_assignments == false) {
+        dummy_ecdsa_constraint(composer, input);
+    }
 
     auto new_sig = ecdsa_convert_signature(composer, input.signature);
 
@@ -125,6 +131,56 @@ void create_ecdsa_verify_constraints(Composer& composer, const EcdsaSecp256k1Con
                                                                    secp256k1_ct::g1_bigfr_ct>(message, public_key, sig);
     bool_ct signature_result_normalized = signature_result.normalize();
     composer.assert_equal(signature_result_normalized.witness_index, input.result);
+}
+
+// Add dummy constraints for ECDSA because when the verifier creates the
+// constraint system, they usually use zeroes for witness values.
+//
+// This does not work for ECDSA as the signature, r, s and public key need
+// to be valid.
+void dummy_ecdsa_constraint(Composer& composer, EcdsaSecp256k1Constraint const& input)
+{
+
+    std::vector<uint32_t> pub_x_indices_;
+    std::vector<uint32_t> pub_y_indices_;
+    std::vector<uint32_t> signature_;
+    signature_.resize(64);
+
+    // Create a valid signature with a valid public key
+    crypto::ecdsa::key_pair<secp256k1_ct::fr, secp256k1_ct::g1> account;
+    account.private_key = 10;
+    account.public_key = secp256k1_ct::g1::one * account.private_key;
+    uint256_t pub_x_value = account.public_key.x;
+    uint256_t pub_y_value = account.public_key.y;
+    std::string message_string = "Instructions unclear, ask again later.";
+    crypto::ecdsa::signature signature =
+        crypto::ecdsa::construct_signature<Sha256Hasher, secp256k1_ct::fq, secp256k1_ct::fr, secp256k1_ct::g1>(
+            message_string, account);
+
+    // Create new variables which will reference the valid public key and signature.
+    // We don't use them in a gate, so when we call assert_equal, they will be
+    // replaced as if they never existed.
+    for (size_t i = 0; i < 32; ++i) {
+        uint32_t x_wit = composer.add_variable(pub_x_value.slice(248 - i * 8, 256 - i * 8));
+        uint32_t y_wit = composer.add_variable(pub_y_value.slice(248 - i * 8, 256 - i * 8));
+        uint32_t r_wit = composer.add_variable(signature.r[i]);
+        uint32_t s_wit = composer.add_variable(signature.s[i]);
+        pub_x_indices_.emplace_back(x_wit);
+        pub_y_indices_.emplace_back(y_wit);
+        signature_[i] = r_wit;
+        signature_[i + 32] = s_wit;
+    }
+
+    // Call assert_equal(from, to) to replace the value in `to` by the value in `from`
+    for (size_t i = 0; i < input.pub_x_indices.size(); ++i) {
+        composer.assert_equal(pub_x_indices_[i], input.pub_x_indices[i]);
+    }
+    for (size_t i = 0; i < input.pub_y_indices.size(); ++i) {
+        composer.assert_equal(pub_y_indices_[i], input.pub_y_indices[i]);
+    }
+    for (size_t i = 0; i < input.signature.size(); ++i) {
+        composer.assert_equal(signature_[i], input.signature[i]);
+    }
 }
 
 } // namespace acir_format
